news-data-pipeline-GCP/weekly-fetcher/requirements.txt
requests
google-cloud-storage
functions-framework
flask
openai
#python-dotenv -- For local .env